ec2/requirements.txt
asgiref==3.4.1
Django==3.2.8
gunicorn==20.1.0
numpy==1.21.3
opencv-python-headless==4.5.4.58
Pillow==8.4.0
psycopg2-binary==2.9.1
pytesseract==0.3.8
pytz==2021.3
sqlparse==0.4.2
torch==1.10.0+cpu
torchaudio==0.10.0+cpu
torchvision==0.11.1+cpu
typing-extensions==3.10.0.2
